requirement.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
matplotlib>=3.8.0
groq>=0.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
werkzeug>=3.0.0